Python-Algorithm/music-spotify-AI-Algo/requirements.txt
fastapi==0.75.2
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.0
pymongo==4.1.1
pyparsing==3.0.8
python-dotenv==0.20.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sklearn==0.0
typing==3.7.4.3
uvicorn==0.17.6
